requirements.txt
# WSP2AGENT pinned dependencies
serpapi==0.1.5
lxml>=4.9.0
beautifulsoup4>=4.13.0
requests>=2.31.0
pandas>=2.0.0
reportlab>=3.6.0
google-api-python-client>=2.70.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0
streamlit>=1.28.0
selenium>=4.15.0
pytest>=7.4.0
pytest-cov>=4.1.0
